LayoutTests/http/tests/security/video-cross-origin-accesssameorigin-expected.txt

EVENT(playing)
END OF TEST
